+++ uno/solvers/QP/BQPDSolver.cpp
@@ -31,11 +31,13 @@ BQPDSolver::BQPDSolver(size_t number_variables, size_t number_constraints, size_
       w_(n_ + m_), gradient_solution_(n_), residuals_(n_ + m_), e_(n_ + m_),
       size_hessian_sparsity_(quadratic_programming ? maximum_number_nonzeros + n_ + 3 : 0),
       size_hessian_workspace_(maximum_number_nonzeros_ + kmax_ * (kmax_ + 9) / 2 + 2 * n_ + m_ + mxwk0_),
-      size_hessian_sparsity_workspace_(size_hessian_sparsity_ + kmax_ + mxiwk0_), hessian_(size_hessian_workspace_),
-      hessian_sparsity_(size_hessian_sparsity_workspace_), solution(number_variables) {
+      size_hessian_sparsity_workspace_(size_hessian_sparsity_ + kmax_ + mxiwk0_),
+      hessian_(size_hessian_workspace_),
+      hessian_sparsity_(size_hessian_sparsity_workspace_),
+      solution(number_variables) {
    // active set
    for (size_t i = 0; i < this->n_ + this->m_; i++) {
-      this->ls_[i] = i + this->fortran_shift;
+      this->ls_[i] = (int) (i + this->fortran_shift);
    }
 }
 
@@ -46,20 +48,19 @@ Direction BQPDSolver::solve_QP(const std::vector<Range>& variables_bounds, const
       this->hessian_[i] = hessian.matrix[i];
    }
    /* Hessian sparsity */
-   this->hessian_sparsity_[0] = hessian.number_nonzeros + 1;
+   this->hessian_sparsity_[0] = (int) (hessian.number_nonzeros + 1);
    for (size_t i = 0; i < hessian.number_nonzeros; i++) {
-      this->hessian_sparsity_[i + 1] = hessian.row_index[i] + this->fortran_shift;
+      this->hessian_sparsity_[i + 1] = (int) (hessian.row_index[i] + this->fortran_shift);
    }
    for (size_t i = 0; i < hessian.dimension + 1; i++) {
-      this->hessian_sparsity_[hessian.number_nonzeros + i + 1] =
-            hessian.column_start[i] + this->fortran_shift;
+      this->hessian_sparsity_[hessian.number_nonzeros + i + 1] = (int) (hessian.column_start[i] + this->fortran_shift);
    }
 
    // if extra variables have been introduced, correct hessian.column_start
-   int i = hessian.number_nonzeros + hessian.dimension + 2;
-   int last_value = hessian.column_start[hessian.dimension];
+   size_t i = hessian.number_nonzeros + hessian.dimension + 2;
+   size_t last_value = hessian.column_start[hessian.dimension];
    for (size_t j = hessian.dimension; j < this->n_; j++) {
-      this->hessian_sparsity_[i] = last_value + this->fortran_shift;
+      this->hessian_sparsity_[i] = (int) (last_value + this->fortran_shift);
       i++;
    }
 
@@ -76,10 +77,10 @@ Direction BQPDSolver::solve_subproblem(const std::vector<Range>& variables_bound
       const SparseVector<double>& linear_objective, const std::vector<SparseVector<double>>& constraints_jacobian, const std::vector<double>& x) {
    /* initialize wsc_ common block (Hessian & workspace for bqpd) */
    // setting the common block here ensures that several instances of BQPD can run simultaneously
-   wsc_.kk = this->maximum_number_nonzeros_;
-   wsc_.ll = this->size_hessian_sparsity_;
-   wsc_.mxws = this->size_hessian_workspace_;
-   wsc_.mxlws = this->size_hessian_sparsity_workspace_;
+   wsc_.kk = (int) this->maximum_number_nonzeros_;
+   wsc_.ll = (int) this->size_hessian_sparsity_;
+   wsc_.mxws = (int) this->size_hessian_workspace_;
+   wsc_.mxlws = (int) this->size_hessian_sparsity_workspace_;
    kktalphac_.alpha = 0; // inertia control
 
    DEBUG << "objective gradient: ";
@@ -98,28 +99,28 @@ Direction BQPDSolver::solve_subproblem(const std::vector<Range>& variables_bound
    int current_index = 0;
    linear_objective.for_each([&](size_t i, double derivative) {
       this->jacobian_[current_index] = derivative;
-      this->jacobian_sparsity_[current_index + 1] = i + this->fortran_shift;
+      this->jacobian_sparsity_[current_index + 1] = (int) (i + this->fortran_shift);
       current_index++;
    });
    for (size_t j = 0; j < this->m_; j++) {
       constraints_jacobian[j].for_each([&](size_t i, double derivative) {
          this->jacobian_[current_index] = derivative;
-         this->jacobian_sparsity_[current_index + 1] = i + this->fortran_shift;
+         this->jacobian_sparsity_[current_index + 1] = (int) (i + this->fortran_shift);
          current_index++;
       });
    }
    current_index++;
    this->jacobian_sparsity_[0] = current_index;
    // header
-   int size = 1;
-   this->jacobian_sparsity_[current_index] = size;
+   size_t size = 1;
+   this->jacobian_sparsity_[current_index] = (int) size;
    current_index++;
    size += linear_objective.size();
-   this->jacobian_sparsity_[current_index] = size;
+   this->jacobian_sparsity_[current_index] = (int) size;
    current_index++;
    for (size_t j = 0; j < this->m_; j++) {
       size += constraints_jacobian[j].size();
-      this->jacobian_sparsity_[current_index] = size;
+      this->jacobian_sparsity_[current_index] = (int) size;
       current_index++;
    }
 
@@ -181,7 +182,7 @@ Direction BQPDSolver::generate_direction() {
       }
       else {
          // general constraint
-         int constraint_index = index - this->n_;
+         size_t constraint_index = index - this->n_;
          direction.constraint_partition.feasible.push_back(constraint_index);
          direction.constraint_partition.constraint_feasibility[constraint_index] = FEASIBLE;
          if (0 <= this->ls_[j]) { /* lower bound active */
@@ -200,7 +201,7 @@ Direction BQPDSolver::generate_direction() {
       size_t index = std::abs(this->ls_[j]) - this->fortran_shift;
 
       if (this->n_ <= index) { // general constraints
-         int constraint_index = index - this->n_;
+         size_t constraint_index = index - this->n_;
          if (this->residuals_[index] < 0.) { // infeasible constraint
             direction.constraint_partition.infeasible.push_back(constraint_index);
             if (this->ls_[j] < 0) { // upper bound violated
